requirements.txt
QuizEngine4Trivia-TirsvadCLI>=0.0.6
